[refactor] AuthService 인터페이스 작성 및 메서드 설명 추가
