+++ src/main/java/com/example/spot/security/utils/SecurityUtils.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 public class SecurityUtils {
 
+    // 현재 인증된 사용자의 ID를 반환
     public static Long getCurrentUserId() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication == null || !authentication.isAuthenticated()) {
@@ -16,6 +17,7 @@ public class SecurityUtils {
         return Long.valueOf(authentication.getName());
     }
 
+    // 현재 인증된 사용자의 ID와 매개변수로 전달된 ID가 일치하는지 확인
     public static void verifyUserId(Long memberId) {
         Long currentUserId = getCurrentUserId();
         if (!Objects.equals(currentUserId, memberId)) {
